+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 
+ * [Added] Find if of closest points in the dataframe
  * [Added] `kwargs` can be used in `imsave()`
  * [Added] Flatten keys in struct. The list and ndarray can be converted to simple cols.